feat: 정보게시판 상세페이지 api 연동 1차
